refactor: keep lib/favicon framework-free and move its hook
"use client" at module scope made the pure getFaviconUrl a client reference
too, so a server component importing it would fail at call time for no
visible reason. The hook moves to lib/hooks/ where every other hook in the
repo lives; the registry still ships both files with citations because the
bundler follows the import.
